backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv==1.0.1
numpy==2.1.2
